properly refactor testing setup into lib
