+++ kernel/src/testsw/writer.rsw
@@ -1,0 +1,19 @@
+#![no_std]
+#![no_main]
+#![feature(custom_test_frameworks)]
+#![test_runner(hexium_os::test_runner)]
+#![reexport_test_harness_main="test_main"]
+
+use core::panic::PanicInfo;
+
+#[unsafe(no_mangle)]
+unsafe extern "C" fn kmain() -> ! {
+  test_main();
+  loop{}
+}
+
+#[panic_handler]
+fn panic(info: &PanicInfo) -> ! {
+  hexium_os::test_panic_handler(info);
+  loop{}
+}
